requirements.txt
google-adk==1.27.4
python-dotenv==1.2.2
google-adk[extensions]

